requirements-dev.txt
pytest
pytest-django
pdbpp

# Release tools
setuptools
twine
wheel
